Renamed all instance of Arrav to RagePS removed the use of RestfulNexus rest api for forum interactions created and tested queries for getting, creating, updating last active forums added queries for checking if exists and loading/saving character files implemented properties from environment and added debug property

+++ src/main/java/com/rageps/world/entity/actor/player/PlayerCredentials.java
@@ -15,7 +15,7 @@ public final class PlayerCredentials {
 	public long usernameHash;
 	public String username;
 	public String password;
-	public long databaseId;
+	public int databaseId;
 	private PlayerEmail playerEmail;
 
 	public PlayerCredentials(String username, String password) {
